+++ sketch/src/main/java/com/github/panpf/sketch/decode/WebpAnimatedDrawableDecoder.kt
@@ -58,10 +58,9 @@ fun ComponentRegistry.Builder.supportAnimatedWebp(): ComponentRegistry.Builder =
  */
 @RequiresApi(Build.VERSION_CODES.P)
 class WebpAnimatedDrawableDecoder(
-    sketch: Sketch,
     requestContext: RequestContext,
     dataSource: DataSource,
-) : BaseAnimatedImageDrawableDecoder(sketch, requestContext, dataSource) {
+) : BaseAnimatedImageDrawableDecoder(requestContext, dataSource) {
 
     class Factory : DrawableDecoder.Factory {
 
@@ -82,7 +81,7 @@ class WebpAnimatedDrawableDecoder(
             ) {
                 val imageFormat = ImageFormat.parseMimeType(fetchResult.mimeType)
                 if ((imageFormat == null || imageFormat == ImageFormat.WEBP) && fetchResult.headerBytes.isAnimatedWebP()) {
-                    return WebpAnimatedDrawableDecoder(sketch, requestContext, dataSource)
+                    return WebpAnimatedDrawableDecoder(requestContext, dataSource)
                 }
             }
             return null