+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### sketch
 
 * improve: Compatible with still images encoded using WebP Animated panpf Moments ago [#179](https://github.com/panpf/sketch/issues/179)
+* improve: Improved memory caching, now all BitmapDrawable is converted to SketchCountBitmapDrawable in MemoryCacheRequestInterceptor and cached in memory, so all BitmapDecoders can return BitmapDrawable directly
 
 # v3.2.1-beta02